Allocate float_buffer_ in the initializer list.
This may fix a Dr. Memory error: "allocated with operator new, freed
with operator delete[]". I suspect this is a false positive; in the
existing implementation the reset causes a delete[] on NULL. This is
a no-op of course, but Dr. Memory might be flagging it. We shall see.

In any case, this change is an improvement.

BUG=2321
TBR=bjornv

Review URL: https://webrtc-codereview.appspot.com/2215004