Fixing broken tests in voe_auto_test extended
This CL fixes the problem with voe_auto_test extended-codec test, as well as
extended-file test. First problem was that Opus was not added as a special case, like the other codecs, and the second problem was that the tests were not updated when test files were moved to the resources catalogue.

There are still some tests that fails. Here is a list of all extended tests and their status:

Base: fails - the reason seem to be that external transport has been removed.
CallReport: passes
Codec: passes (with this CL)
DTMF: passes
Encryption: fails or is dissabled?
VoEExternalMedia: passes
File: passes (with this CL)
Hardware: passes
NetEqStats: empty?
Network: passes
RTP_RTCP: fails
VideoSync: fails
VolumeControl: passes

BUG=issue2234
R=andrew@webrtc.org, henrika@webrtc.org, xians@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/2023004